Fix for requeued job in slurm.
Jobs that are requeued for some reason (preemptive jobs, node failure, etc)
will create two records in the slurm accounting database but only one will show up.
Depending on the frequency of bart the first or the second part of the job will end
up in the sgas database. 

For this to work with sgas idtimestamp must be enabled.